+++ generate-readme-and-changelog.sh
@@ -1,0 +1,4 @@
+#!/bin/sh
+
+sphinx-build -q -b rst -d build/readme docs/readme .
+sphinx-build -q -b rst -d build/changelog docs/changelog .