+++ src/sphinx_rst_builder/__init__.py
@@ -1,27 +1,20 @@
 # -*- coding: utf-8 -*-
+# SPDX-License-Identifier: BSD-2-Clause
+"""Sphinx extension to output reST files.
+
 """
-    sphinxcontrib.restbuilder
-    =========================
-
-    Sphinx extension to output reST files.
-
-    .. moduleauthor:: Freek Dijkstra <software@macfreek.nl>
-
-    :copyright: Copyright 2012-2013 by Freek Dijkstra.
-    :license: BSD, see LICENSE.txt for details.
-"""
-
-from __future__ import (print_function, unicode_literals, absolute_import)
 
 from sphinx.builders import Builder
 from sphinx.writers.text import STDINDENT
-from .builders.rst import RstBuilder
-from .writers.rst import RstWriter
+
+from ._builder import RstBuilder
+from ._writer import RstWriter
 
 
 
 def setup(app):
-    app.require_sphinx('1.0')
+    """Setup this extension"""
+    app.require_sphinx('2.0')
     app.add_builder(RstBuilder)
     app.add_config_value('rst_file_suffix', ".rst", False)
     """This is the file name suffix for reST files"""